Dotnet Corrections

+++ Jenkins/PipelineScripts/Pipeline-Dotnet.groovy
@@ -2,57 +2,77 @@ import groovy.json.JsonSlurperClassic
 def modules = [:]
 pipeline {
     agent any
-    environment {
-        port =
-        Code_Repo_URL = 
-        SAST_Server_IP = 
-        Sonar_Port = 
-        Orchestrator_POST_URL = 
-        Sonar_Token = 
+    environment 
+    {
+        
+        branches = 'develop,master' //List of valid branches
+
+        Code_Repo_URL = 'https://LeonardoMarazzo@bitbucket.org/directvla/dtvweb.git'
+        
+        SAST_Server_IP = '192.168.0.238'
+        SAST_Server_User = 'maxpowersi'
+        SAST_Server_Repository_SAST_Path = '/home/maxpowersi/TheFantasticSecDevOps/SAST'
+        SAST_Server_SSH_Port = 4222
+        
+        Sonar_Token = ''
+        Sonar_Port = 9000
+        
+        Orchestrator_POST_URL = 'https://726b58897291.ngrok.io/add_code_vulnerability/'
+        Orchestrator_START_URL = 'https://726b58897291.ngrok.io/start'
+        Orchestrator_END_URL = 'https://726b58897291.ngrok.io/end'
+
     }
+
     stages {
         stage('Import scripts files from Git'){
             steps{
                 script{
                     try {
+
+                        //Check if trigred branch is a valid branch.
+                        if(!(env.branches.split(',').contains(env.branch))) {
+                            SkipBuild = 'YES'
+                            print(SkipBuild)
+                        }
+                        if (SkipBuild == 'YES'){
+                            currentBuild.result = 'SUCCESS'
+                            return
+                        }
+
                         sh "rm -rf \$(pwd)/*"
         
                         //Importings scripts from gitlab
                         git credentialsId: 'git-secpipeline-token', url: 'https://github.com/badBounty/TheFantasticSecDevOps.git'
 
+                        modules.Notifier = load "Jenkins/PipelineScripts/Notifier.groovy"
+                        modules.Notifier_Slack = load "Jenkins/PipelineScripts/Notifier-Slack.groovy"
+
+                        modules.Notifier.sendMessage('','good','Stage: "Import-Jenkins-Scripts": INIT')
+
+                        modules.Notifier.Init(modules.Notifier_Slack)
+
                         //Load sripts in collection
                         modules.Install_GitCheckout = load "Jenkins/PipelineScripts/Install-GitCheckout.groovy"
                         modules.Install_Dependecies = load "Jenkins/PipelineScripts/Install-DotNetDependecies.groovy"
-
                         modules.SAST_Deployment = load "Jenkins/PipelineScripts/SAST-Deployment.groovy"
-
                         modules.SAST_Sonarqube = load "Jenkins/PipelineScripts/SAST-SonarQube-Dotnet.groovy"
                         modules.SAST_SonarResults = load "Jenkins/PipelineScripts/SAST-SonarResults.groovy"
                         modules.SAST_DotNet = load "Jenkins/PipelineScripts/SAST-Dotnet.groovy"
-
                         modules.SAST_Destroy = load "Jenkins/PipelineScripts/SAST-Destroy.groovy"
+                        modules.SAST_PostResults = load "Jenkins/PipelineScripts/SAST-PostResults.groovy"
+                        modules.SAST_SendVulnsLog = load "Jenkins/PipelineScripts/SAST-SendVulnsLog.groovy"
                         
-                        //modules.Build_Dotnet = load "/var/jenkins_home/PipelineScripts/Build-Dotnet.groovy"
-                        //modules.Build_DockerBuild = load "/var/jenkins_home/PipelineScripts/Build-DockerBuild.groovy"
                         
-                        //modules.Deploy_DockerRun = load "/var/jenkins_home/PipelineScripts/Deploy-DockerRun.groovy"
 
-                        modules.Notifier = load "Jenkins/PipelineScripts/Notifier.groovy"
-                        modules.Notifier_Slack = load "Jenkins/PipelineScripts/Notifier-Slack.groovy"
-                        
-                        modules.Notifier.Init(modules.Notifier_Slack)
-                        modules.Notifier.sendMessage('','good','Pulling script files from github') 
-                        modules.Notifier.sendMessage('','good','Git Pulling: SUCCESS')
-                        
-                        print('------Stage "Import scripts files from Git": SUCCESS ------')
-                    } catch(Exception e) {
-
-                        //print(e.printStackTrace())
-                        currentBuild.result = 'FAILURE'      
-                        modules.Notifier.sendMessage('','danger','An error occurred in the "Import scripts files from Git" stage') 
-                        modules.Notifier.sendMessage('','danger',"Git Pulling: FAILURE")
-
-                        print('------Stage "Import scripts files from Git": FAILURE ------')
+                        modules.Notifier.sendMessage('','good','Stage: "Import-Jenkins-Scripts": SUCCESS')
+                        print('Stage: "Import-Jenkins-Scripts": SUCCESS')
+                        print(modules)
+                    }
+                    catch(Exception e)
+                    {
+                        print(modules)
+                        currentBuild.result = 'FAILURE'
+                        print('Stage: "Import-Jenkins-Scripts": FAILURE')
                     } // try-catch-finally
                 } // script
             } // steps
@@ -61,6 +81,10 @@ pipeline {
         stage('Install-GitCheckout'){
             steps{
                 script{
+                    if (SkipBuild == 'YES'){
+                        currentBuild.result = 'SUCCESS'
+                        return
+                    }
                     modules.Install_GitCheckout.runStage()
                 }
             }
@@ -69,6 +93,10 @@ pipeline {
         stage('Install-Dependencies'){
             steps{
                 script{
+                    if (SkipBuild == 'YES'){
+                        currentBuild.result = 'SUCCESS'
+                        return
+                    }
                     modules.Install_Dependecies.runStage()
                 }
             }
@@ -77,6 +105,10 @@ pipeline {
         stage('SAST-Deployment'){
             steps{
                 script{
+                    if (SkipBuild == 'YES'){
+                        currentBuild.result = 'SUCCESS'
+                        return
+                    }
                     modules.SAST_Deployment.runStage()
                 }
             }
@@ -85,16 +117,23 @@ pipeline {
         stage('SAST-SonarQube'){
             steps{
                 script{
+                    if (SkipBuild == 'YES'){
+                        currentBuild.result = 'SUCCESS'
+                        return
+                    }
                    modules.SAST_SonarQube_DotNet.runStage()
                 }
             }
         }
 
-        stage('SAST-Dotnet#'){
+        stage('SAST-DotnetCore'){
             steps{
                 script{
+                    if (SkipBuild == 'YES'){
+                        currentBuild.result = 'SUCCESS'
+                        return
+                    }
                    modules.SAST_DotNet.runStage()
-                   modules.SAST_DotNet.parseVulns()
                 }
             }
         }
@@ -102,6 +141,10 @@ pipeline {
         stage('SAST-SonarResults'){
             steps{
                 script{
+                    if (SkipBuild == 'YES'){
+                        currentBuild.result = 'SUCCESS'
+                        return
+                    }
                     modules.SAST_SonarResults.runStage()
                 }
             }
@@ -110,12 +153,47 @@ pipeline {
         stage('SAST-Destroy'){
             steps{
                 script{
+                    if (SkipBuild == 'YES'){
+                        currentBuild.result = 'SUCCESS'
+                        return
+                    }
                     modules.SAST_Destroy.runStage()
                 }
             }
         }
 
-        stage('Build'){
+        stage('SAST-PostResults')
+        {
+            steps
+            {
+                script
+                {
+                    if (SkipBuild == 'YES'){
+                        currentBuild.result = 'SUCCESS'
+                        return
+                    }
+                    modules.SAST_PostResults.runStage(modules.Notifier, vulns)
+
+                }
+            }
+        }
+
+        stage('SAST-SendVulnsLog')
+        {
+            steps
+            {
+                script
+                {
+                    if (SkipBuild == 'YES'){
+                        currentBuild.result = 'SUCCESS'
+                        return
+                    }
+                    modules.SAST_SendVulnsLog.runStage(modules.Notifier)
+                }
+            }
+        }
+
+        /*stage('Build'){
             steps{
                 script{
                     modules.Build_Dotnet.runStage()
@@ -123,7 +201,7 @@ pipeline {
             }
         }
         
-        /*stage('Build-DockerBuild'){
+        stage('Build-DockerBuild'){
             steps{
                 script{
                     modules.Build_DockerBuild.runStage()
